Change load timeout type to Double

+++ adapty/src/commonMain/kotlin/com/adapty/kmp/internal/plugin/request/AdaptyGetOnboardingRequest.kt
@@ -8,7 +8,7 @@ internal data class AdaptyGetOnboardingRequest(
     @SerialName("placement_id") val placementId: String,
     @SerialName("locale") val locale: String? = null,
     @SerialName("fetch_policy") val fetchPolicy: AdaptyPaywallFetchPolicyRequest? = null,
-    @SerialName("load_timeout") val loadTimeoutInSeconds: Long? = null,
+    @SerialName("load_timeout") val loadTimeoutInSeconds: Double? = null,
 )
